requirements.txt
streamlit
google-generativeai
praw
requests
wikipedia
python-dotenv